moved command to new file
